+++ examples/chars.rs
@@ -1,20 +1,19 @@
 use std::io::{self, stdout, Read, Write};
 
 use termal::{
-    codes,
     error::Result,
     raw::{disable_raw_mode, enable_raw_mode},
 };
 
 fn main() -> Result<()> {
-    print!("{}{}\x1b[?1004h", codes::REQUEST_DEVICE_ATTRIBUTES, "");
+    print!("\x1b[3q");
     _ = stdout().flush();
     enable_raw_mode()?;
 
     start()?;
 
     disable_raw_mode()?;
-    print!("{}\x1b[?1004l", "");
+    print!("");
     _ = stdout().flush();
 
     Ok(())