+++ termal_core/src/raw/events/mod.rs
@@ -1,9 +1,8 @@
 mod csi;
 mod event;
 mod key;
-mod mouse;
+pub mod mouse;
+mod status;
 mod term_attr;
 
-pub use event::*;
-pub use key::*;
-pub use term_attr::*;
+pub use self::{event::*, key::*, status::*, term_attr::*};