+++ termal_core/src/raw/events/status.rs
@@ -1,0 +1,8 @@
+use super::TermAttr;
+
+#[derive(Debug, Clone, Copy)]
+pub enum Status {
+    Attributes(TermAttr),
+    Ok,
+    CursorPosition { x: usize, y: usize },
+}