+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 - Move around some logic around sixels.
 - Add texels - drawing images using charactes.
 - Add option for nearest when scaling image.
-- Add codes related to mouse events
-- Parse mouse events (classic, SGR, UTF-8, URXVT)
-- Parse focus events
+- Add new codes.
+- Parse mouse events (classic, SGR, UTF-8, URXVT).
+- Parse focus events.
+- Parse cursor position responses.
 
 ## v1.2.2
 - Fix `writemcln` macro.